requirements.txt
asgiref==3.2.10
astroid==2.4.2
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.1
django-mapbox-location-field==1.6.3
djangorestframework==3.11.1
geojson==2.5.0
gunicorn==20.0.4
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
static3==0.7.0
toml==0.10.1
whitenoise==5.2.0
wrapt==1.12.1
